requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
mysqlclient==2.2.4
python-dotenv==1.0.1